Position-Based Fluid (almost) working!
This is includes the main PBF code, which a more modern approach to
simulating fluids.

And also a new precompute pass that places all particles on a pixel grid
so that we can efficiently find neighbouring particles to run the main
kernel code on. Reducing the complexity from O(n^2) to O(n * radius)

This commit is working, it just lacks the application of the kernel on
the neighbouring particles. So the simulation isn't accurate.

Also I want to do a lot of refactoring.