Propogating lambda to neighbouring particles
There's a lambda term that's generated by each particle, that then has
to be applied to each neighbouring particle. I'm currently doing this by
applying the lambda on the *subsequent* compute pass, I'm not 100% sure
that it's mathemtical.

Also implemented adjustable resolution for the pixel map in which
neighbours are stored.