Grid-neighbour search and algo tweaks for GPU
This commit has a lot of exploration in it, I really fought to
understand how to implement the PBD algo using parellelism.

To be 100% sure the neighbour search wasn't causing problems, I changed
it from pixel-based to grid-based. But I think it was ok all along.

The deeper problem was how to propogate the lambda/pressure/forces from
a particle out to its neighbours. On the CPU that's trivial, but in
parallel on a GPU it's a bit trickier.

The simulation still doesn't exactly match @cerrno's CPU-based PBD
version at https://github.com/cerrno/pbd-fluid-rs The main problem seems
to be the sudden propogation of the forces, it looks like shock waves
are flowing through the liquid. But interestingly they don't appear if
the time step is increased.

My theory is that the sequential propogation of the CPU algo eases the
propogation of forces because of its random, additive, piecemeal nature.