Moved to Bevy! First working code (no physics yet)
I was thinking of doing this for a while, but what got me to commit was
finding https://github.com/AnthonyTornetta/bevy_easy_compute that gives
a very easy interface for coding GPU workloads in Bevy.

This commit completely rewrites everything 😲 And it doesn't even do
proper physics yet, but everything is in place for that.

I'm planning to move away from SPH physics to PIC/FLIP physics. Using
Ten Minute Physics' and @d;i's examples [1,2]

This is also the first use of a dedicated compiler for the Rust shaders:
https://github.com/tombh/rust-gpu-compiler

1. https://matthias-research.github.io/pages/tenMinutePhysics/18-flip.html
2. http://david.li/fluid/