Spatial binning (with some temporary hacks)
Spatial binning, or spatial hashing as it's more commonly known is an
efficient data structure for speeding up nearest neighbour search.

See: https://matthias-research.github.io/pages/tenMinutePhysics/11-hashing.pdf

I don't actually do any particle searches here, but I've got the CPU
generating and storing the spatial binning format. And the GPU is
consuming this format. The main next step will be to get the GPU to do
prefix summing so that it can produce its own spatial bins.

The main hack at the moment is that the CPU is recomputing the spatial
bins and uploading them to the GPU for every frame. This is definitely
inefficient hopefully the next commit fixes this.