GPU-calculated partial sum spatial bins
The spatial bin data structure is now calculated on the GPU.

It turned out harder than I'd hoped because I don't currently have a
machine that supports GPU "subgroups". It looks like the state of the
art has arrived at an algorithm that uses these subgroups, so most
references and examples on the internet are focussed on that. But my
current driver (Asahi OpenGL 4.6) doesn't seem to support them. On the
plus side what I've implemented is probably supported on more hardware.

TODO:
  - "You're A Pixel" example can't control the red pixel. I think it
    will require adding pixel IDs so that particles can be tracked
    between the CPU and GPU.
  - Use a more modern, subgroup-based prefix sum algorithm, see:
    https://github.com/b0nes164/GPUPrefixSums
  - Replace most of the WGSL shaders with SPIR-V once rust-gpu supports
    atomics.